Alerts aggregates

+++ Ganga/client/src/Components/Panels/AdminPanel/vendorProducts.jsx
@@ -4,6 +4,7 @@ import { DataGrid } from "@mui/x-data-grid";
 import { useDispatch, useSelector } from "react-redux";
 import { getProduct } from "../../Redux/Actions/actions";
 import { TiDeleteOutline } from "react-icons/ti";
+import { Swal } from 'sweetalert2';
 import { BsPencilSquare } from "react-icons/bs";
 
 import { deleteProduct } from "../../Redux/Actions/actions";
@@ -54,8 +55,26 @@ export default function VendorProducts({ id }) {
   ];
 
   function handleDelete(id) {
-    dispatch(deleteProduct(id));
-    setRows(rows.filter((i) => i.id !== id));
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Se borraran todos los datos del producto.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Confirmar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Eliminado!',
+          'El producto ha sido eliminado.',
+          'success'
+          )
+          dispatch(deleteProduct(id));
+          setRows(rows.filter((i) => i.id !== id));
+      }
+    })
     window.location.reload();
   }
